[CHORE/VG-28] InputAuto 함수 파일 분리
