+++ Source/GA6thFinal_Framework/GameEngine/GameEngine.vcxproj.filters
@@ -564,6 +564,9 @@
     <ClInclude Include="Source\Engine\ClassCore\Command\InputAutoCommand.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
+    <ClInclude Include="Source\Engine\ClassCore\InputAuto.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Source\pch.cpp">
