+++ Source/GA6thFinal_Framework/GameEngine/Source/Engine/ClassCore/Command/InputAutoCommand.h
@@ -34,7 +34,7 @@ namespace Command
 
             virtual void Undo()
             {
-                 *_pData = _prev;
+                *_pData = _prev;
             }
         };
