+++ Source/GA6thFinal_Framework/GameEngine/Source/Engine/ClassCore/ReflectHelper.h
@@ -246,6 +246,7 @@ protected:                                                                      
         ImGui::PopID();                                                                                     \
     }
 
+
 namespace ReflectHelper
 {
     namespace ImGuiDraw
@@ -257,8 +258,7 @@ namespace ReflectHelper
                 int                 step      = 0;
                 int                 step_fast = 0;
                 ImGuiInputTextFlags flags     = 0;
-            }
-            _int;
+            } _int;
 
             struct Float
             {
@@ -266,8 +266,7 @@ namespace ReflectHelper
                 float               step_fast = 0;
                 std::string         format    = "%.3f";
                 ImGuiInputTextFlags flags     = 0;
-            }
-            _float;
+            } _float;
 
             struct Double
             {
@@ -275,16 +274,14 @@ namespace ReflectHelper
                 double              step_fast = 0;
                 std::string         format    = "%.6f";
                 ImGuiInputTextFlags flags     = 0;
-            }
-            _double;
+            } _double;
 
             struct String
             {
                 ImGuiInputTextFlags    flags     = 0;
                 ImGuiInputTextCallback callback  = nullptr;
                 void*                  user_data = nullptr;
-            }
-            _string;
+            } _string;
 
             struct Vector2
             {
@@ -293,8 +290,7 @@ namespace ReflectHelper
                 float            v_max   = 0.f;
                 std::string      format  = "%.3f";
                 ImGuiSliderFlags flags   = 0;
-            }
-            _Vector2;
+            } _Vector2;
 
             struct Vector3
             {
@@ -303,8 +299,7 @@ namespace ReflectHelper
                 float            v_max   = 0.f;
                 std::string      format  = "%.3f";
                 ImGuiSliderFlags flags   = 0;
-            }
-            _Vector3;
+            } _Vector3;
 
             struct Vector4
             {
@@ -313,362 +308,19 @@ namespace ReflectHelper
                 float            v_max   = 0.f;
                 std::string      format  = "%.3f";
                 ImGuiSliderFlags flags   = 0;
-            }
-            _Vector4;
+            } _Vector4;
 
-            //InputAuto return 직전에 호출해주는 함수객체 입니다. 
-            //매개변수로 Input 여부랑 접근한 맴버의 이름을 전달해줍니다.
-            //ImGuiDrawPropertys() 함수 호출 이후 NULL로 초기화됩니다.
+            // InputAuto return 직전에 호출해주는 함수객체 입니다.
+            // 매개변수로 Input 여부랑 접근한 맴버의 이름을 전달해줍니다.
+            // ImGuiDrawPropertys() 함수 호출 이후 NULL로 초기화됩니다.
             std::function<void(bool, std::string_view)> InputEndEvent;
         };
     } // namespace ImGuiDraw
 
-    // serialized helper
     namespace json
     {
-        template <typename Type>
-        std::string SerializedObjet(Type& obj);
-
-        template <typename Type>
-        bool DeserializedObjet(Type& obj, std::string_view data);
-
-        /// <summary>
-        /// <para>전달한 yyjson_val*를 char*형식의 문자열로 변환해서 반환합니다.</para>
-        /// <para>반드시 반환된 값이 nullptr이 아니면 사용후 free로 해제해야 합니다.</para>
-        /// </summary>
-        /// <param name="val :">변환할 val 값</param>
-        /// <returns>변환된 문자열</returns>
         char* yyjsonValToCStr(yyjson_val* val);
-    } // namespace json
-} // namespace ReflectHelper
 
-namespace ReflectHelper
-{
-    // 구현부
-    namespace ImGuiDraw
-    {
-        namespace Private
-        {
-            void EngineLog(int logLevel, std::string_view message, std::source_location location = std::source_location::current());
-
-            template <class T>
-            bool InputAuto(T field, const InputAutoSetting& setting)
-            {
-                using namespace DirectX::SimpleMath;
-                auto NotArrayTypeFunc = [&setting](auto* value, const char* name) {
-                    using value_type = std::remove_cvref_t<decltype(*value)>;
-                    using OriginType = std::remove_cvref_t<
-                        PropertyUtils::get_field_type_t<value_type>>;
-                    constexpr bool isProperty =
-                        PropertyUtils::is_TProperty_v<value_type>;
-                    constexpr bool isSetter = PropertyUtils::is_setter<value_type>;
-                    constexpr bool isLock = isProperty == true && isSetter == false;
-
-                    bool isEdit = false;
-                    bool result = false;
-                    auto& val    = *value;
-
-                    if constexpr (isLock == true)
-                    {
-                        ImGui::BeginDisabled();
-                    }
-
-                    if constexpr (std::is_same_v<OriginType, int>)
-                    {
-                        int input = val;
-                        isEdit    = ImGui::InputInt(name, &input, 
-                            setting._int.step,
-                            setting._int.step_fast,
-                            setting._int.flags);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (std::is_same_v<OriginType, float>)
-                    {
-                        float input = val;
-                        isEdit      = ImGui::InputFloat(
-                            name, &input, 
-                            setting._float.step,
-                            setting._float.step_fast,
-                            setting._float.format.c_str(),
-                            setting._float.flags);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (std::is_same_v<OriginType, double>)
-                    {
-                        double input = val;
-                        isEdit       = ImGui::InputDouble(
-                            name, &input, InputAutoSetting::Double::step,
-                            setting._double.step_fast,
-                            setting._double.format.c_str(),
-                            setting._double.flags);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }                      
-                    }
-                    else if constexpr (std::is_same_v<OriginType, bool>)
-                    {
-                        bool input = val;
-                        isEdit     = ImGui::Checkbox(name, &input);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (std::is_same_v<OriginType, std::string>)
-                    {
-                        static std::string input;
-                        input  = val;
-                        isEdit = ImGui::InputText(
-                            name, &input, 
-                            setting._string.flags,
-                            setting._string.callback,
-                            setting._string.user_data);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (isProperty && std::is_same_v<OriginType, std::string_view>)
-                    {
-                        static std::string input;
-                        input  = val;
-                        isEdit = ImGui::InputText(
-                            name, 
-                            &input, 
-                            setting._string.flags,
-                            setting._string.callback,
-                            setting._string.user_data);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;                               
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (isProperty && std::is_same_v<OriginType, DirectX::SimpleMath::Vector2>)
-                    {
-                        DirectX::SimpleMath::Vector2 input = val;
-                        isEdit = ImGui::DragFloat2(
-                            name, 
-                            &input.x, 
-                            setting._Vector2.v_speed,
-                            setting._Vector2.v_min,
-                            setting._Vector2.v_max,
-                            setting._Vector2.format.c_str(),
-                            setting._Vector2.flags);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (isProperty && std::is_same_v<OriginType, DirectX::SimpleMath::Vector3>)
-                    {
-                        DirectX::SimpleMath::Vector3 input = val;
-                        isEdit = ImGui::DragFloat3(
-                            name, 
-                            &input.x, 
-                            setting._Vector3.v_speed,
-                            setting._Vector3.v_min,
-                            setting._Vector3.v_max,
-                            setting._Vector3.format.c_str(),
-                            setting._Vector3.flags);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else if constexpr (isProperty && std::is_same_v<OriginType, DirectX::SimpleMath::Vector4>)
-                    {
-                        DirectX::SimpleMath::Vector4 input = val;
-                        isEdit = ImGui::DragFloat4(
-                            name, 
-                            &input.x, 
-                            setting._Vector4.v_speed,
-                            setting._Vector4.v_min,
-                            setting._Vector4.v_max,
-                            setting._Vector4.format.c_str(),
-                            setting._Vector4.flags);
-
-                        if constexpr (isProperty == false || isSetter == true)
-                        {
-                            if (isEdit)
-                            {
-                                val = input;
-                            }
-                            if (ImGui::IsItemDeactivatedAfterEdit())
-                            {
-                                result = true;
-                            }
-                        }
-                    }
-                    else
-                    {
-                        EngineLog(
-                            LogLevel::LEVEL_WARNING,
-                            std::format(
-                                "{}, {} - {}", 
-                                (const char*)u8"정의되지 않은 InputAuto 타입입니다.",
-                                typeid(val).name(), 
-                                name
-                                )
-                                .c_str());
-                    }
-                    if constexpr (isLock == true)
-                    {
-                        ImGui::EndDisabled();
-                    }
-                    if constexpr (isProperty == true)
-                    {
-                        if(ImGui::BeginDragDropTarget())
-                        {
-                            val.InvokeDragDropFunc();
-                            ImGui::EndDragDropTarget();
-                        }                   
-                    }
-                    return result;
-                };
-
-                bool isEdit = false;
-                ImGui::PushID(field.value());
-                {
-                    const type_info& type  = typeid(*field.value());
-                    const auto&      name  = field.name();
-                    auto*            value = field.value();
-
-                    using value_type = std::remove_cvref_t<decltype(*value)>;
-                    using OriginType =
-                        PropertyUtils::get_field_type_t<value_type>;
-                    if constexpr (StdHelper::is_std_array_v<OriginType>)
-                    {
-                        if (ImGui::CollapsingHeader((const char*)name.data()))
-                        {
-                            if constexpr (std::ranges::range<decltype(*value)>)
-                            {
-                                int i = 0;
-                                for (auto& val : *value)
-                                {
-                                    isEdit = NotArrayTypeFunc(&val, std::format("[{}]", i).c_str());
-                                    i++;
-                                }
-                            }
-                        }
-                    }
-                    else if constexpr (StdHelper::is_std_vector_v<OriginType>)
-                    {
-                        if constexpr (std::ranges::range<decltype(*value)>)
-                        {
-                            if (ImGui::CollapsingHeader((const char*)name.data()))
-                            {
-                                int  i = 0;
-                                for (auto& val : *value)
-                                {
-                                    isEdit = NotArrayTypeFunc(&val, std::format("[{}]", i).c_str());
-                                    i++;
-                                }
-                                if (ImGui::Button("+"))
-                                {
-                                    value->emplace_back();
-                                    isEdit = true;
-                                }
-                                ImGui::SameLine();
-                                if (ImGui::Button("-"))
-                                {
-                                    if (value->size() > 0)
-                                    {
-                                        value->pop_back();
-                                        isEdit = true;
-                                    }                                 
-                                }
-                            }
-                        }
-                    }
-                    else
-                    {
-                        isEdit = NotArrayTypeFunc(value, name.data());
-                    }
-                }
-
-                if (setting.InputEndEvent)
-                {
-                    setting.InputEndEvent(isEdit, field.name());
-                }
-                ImGui::PopID();
-                return isEdit;
-            }
-        } // namespace Private
-    } // namespace ImGuiDraw
-
-    namespace json
-    {
         template <typename Type>
         inline std::string SerializedObjet(Type& obj)
         {
